refactor: `construct_lookup()` を移動
